Intentando comunicar con la api, para consultar perfil generico.
